fix(test): skip Windows permission tests, close unclosed resources, isolate Linux-specific code
1. Skip permission-related tests on Windows due to OS differences
2. Add missing resource cleanup (logger) in test cases
3. Move Linux-specific code (syscall.Umask) to build-tagged files
